requirements.txt
numpy
pandas
matplotlib
soundfile
praat_parselmouth-0.5.0.dev0-cp313-cp313-manylinux_2_17_x86_64.manylinux2014_x86_64.whl
google-generativeai
pillow
streamlit
streamlit-audiorec
streamlit-advanced-audio
openpyxl
box_sdk_gen
streamlit_authenticator
Authlib
streamlit-js-eval